Payment Formu ayrı Componente böler
